WordContainer을 분리했다
