+++ src/components/FestivalSentence.tsx
@@ -1,38 +1,10 @@
 import styled from 'styled-components';
 
-import { Swiper, SwiperSlide } from 'swiper/react';
-import 'swiper/swiper-bundle.min.css';
-import 'swiper/swiper.min.css';
-import 'swiper/components/navigation/navigation.min.css';
-import 'swiper/components/pagination/pagination.min.css';
-
 import FestivalSentenceItem from './FestivalSentenceItem';
-import useFetchKeyword from '../hooks/useFetchWords';
 import useFetchSentence from '../hooks/useFetchSentence';
+import Word from './Word';
 
 const FestivalSentenceBox = styled.div`
-`;
-
-const WordContainer = styled.div`
-        display:flex;
-        margin-top:12px;
-`;
-
-const Word = styled.div`
-    width: 64px;
-    height: 36px;
-    flex-shrink: 0;
-    border-radius: 33px;
-    background: #F7F7F7;
-    text-align:center;
-    color: #4F4F4F;
-    font-family: SF Pro;
-    font-size: 12px;
-    font-style: normal;
-    font-weight: 600;
-    line-height: 36px;
-    letter-spacing: -0.24px;
-    margin-right:8px;
 `;
 
 const SentenceBox = styled.ul`
@@ -50,7 +22,7 @@ const SentenceBox = styled.ul`
 
 export default function FestivalSentence() {
 
-  const wordList = useFetchKeyword();
+
   const sentences = useFetchSentence();
 
   return (
@@ -59,24 +31,11 @@ export default function FestivalSentence() {
         <SentenceBox>
             { sentences.map((sentence) => (
                 <FestivalSentenceItem sentence={sentence} />
-            ))}
+              ))
+            }
         </SentenceBox>
       </FestivalSentenceBox>
-      <WordContainer>
-        <Swiper
-          spaceBetween={50}
-          slidesPerView={6}
-          allowTouchMove
-          freeMode
-          freeModeMinimumVelocity={0.01}
-        >
-          {wordList.map((word) => (
-            <SwiperSlide key={word.id}>
-              <Word>{word.word}</Word>
-            </SwiperSlide>
-          ))}
-        </Swiper>
-      </WordContainer>
+      <Word />
     </>
   );
 }
